Fixes #4 Remove Hex String Requirement
Replaced most places where the library expects
public and private keys in Hex with a
ReadOnlyMemory<byte> structure instead.  This
refactoring came about as a side affect of debugging
the keys used for creating accounts and then later
used for crypto transfers.

Also removed other uses of ReadOnlySpan due to
the limitations and problematic switching betwen
byte arrays, spans and memory objects.  Now, most
of the public surface leverages the Memory struct
and converts internally to a Span when needed, for
example using the NSec library to import a key from
bytes.