Fixes #12 Context Stack Issues
Modified implementation to align with intent of design
eliminating a bug where setting a property to null resuled
in the same behavior as calling Reset.  The proper behavior
would be to return null, not the value of the parent context.

Additionally made system set default values unresettable to
their intrinsic .net default values via the Reset method.