Fixes #19 Implement Update Account Info
Implemented the crypto update account info method.
Working on this feature triggered a bit of refactoring
and rethinking of some of the API surface and internals.

So that the API calls Create/Update would have a
similar feel, added a CreateAccountParams object to
replace the individual parameters on the create call.
Also removed the default settings from the IContext
object.  This was likely a good choice, as default
 parameters of this kind could have been allowed
to accumulate in the IContext eventually yielding
it a bit too big to be used conveniently.  The
resulting create  parameters object received these
properties instead.  Reasonable defaults to the object
have been added to, so the calling code need not
specify these values on creation.

Note: Proxy staking is not implemented presently
with the test networks, so it is not implemented
in this library at the time.