Fixes #23 Claim Services
Implemented preliminary add/get/remove claim
methods to the hashgraph client.

Note: at the time of writing, there are inconsistent results
returned from the test network when trying to retrieve a claim
attached to an account (that was recently attached, with
confirming receipt)....when this error occurs, the automated
test will be "skipped" because this does not necessarily mean
there is an error in the library, we just can't prove that its
working fully.