Fixes #30 Update Contract
Implemened Update Contract functionality.  Note
the network presently does not actually implement
the expiration time and the change bytecode via
file swap functionality.  The network returns success
but does not change the values.  Therefore, as a precaution
the update parameters for these properties have been
marked as inernal for the time being until they start
working as expected and can be made public.