Copied Contract Source Files into project.
Added the solidity source files into the
project.  At the moment, this is just to
store them in a safe place, at some point
in the future it will be nice to setup a truffle
project that can be injected into the build
pipeline for tests.

Also added an edge case error condition
to the get receipt processing.