Fixes #89 Endorsements Equals Method Bug
Fixed a defect in the Equals implementation when
comparing endorsements of list type that would result
in the comparison returning true for all cases.

Also modified a unit test to reduce the number of
transactions tested because we were seeing random
insufficient TX fees returned for a query that should
not fail (we will address the root cause at some point
in the future).