Fixes #108 Change default renew to 7890000 sec.
The network was changed to only accept 7890000 seconds
as the value accepted for the autorenew period for all entities
including contracts, files and accounts.  Adjusted the defaults
appropriately.

Also fixed other issues caused by network behavorial changes
including tweaking the claims tests, Clames are not yet working
fully so this functionality has yet to be exposed.

The network removed the Get Solidity Address Query from the
system as well.