Fixes #106 Query Fees from Network
Updated the library to ask for the cost of
a query from the network prior to signing
the transaction to pay for the query.  This
reduces the chance of wasting hBars in query
fees.

Introduced a new property on the Query
Contract Call method to set a value to pay
for the size of the payload returned from the
query.  This is a fee in addition to the GAS
necessary to execute the contract method call.