Fixes #105 Support External Signing Code
Introduced a new object to the library called a Signatory.  This object
can hold a private key and can sign a transaction on behalf of the key
owner or it can be given a callback method that can sign a given network
transaction prior to the library submitting it to the network.  This is
now the preferred way to facilitate transaction signing.  This
functionality supersedes the original design decision associated to
place private keys inside the Account object.  The Account object no
longer need be instantiated with any private keys.  Eventually the
Account object will be deprecated and replaced with Address objects in a
future release (presently targeted for removal in version 4.0.0).