Fixes #128 Use Shard.Realm.Num
Throughout the library, we had the reversed
the realm and shard conventions, it has now
been changed to: shard.realm.number.  The order
of parameters in the Address and Gateway objects
has been updated appropriately.

Also updated and added unit tests to increase
test coverage due to gaps identified after the
addicion of the Signatory object to the library.