Fixes #134 Expose Endorsement Properties
Exposed certain internal values of an endorsement
making it possible to retrive the values of public
keys and n of m lists for endorsements retrieved
from GetInfo calls and other scenarios.

This change introduced a small breaking change
by replacing the Endorsement.Type enum with
the (common) KeyType enum.

The KeyType enum
was also widened to include the 'List' value, this
has a minor implication with the Signatory object
as it is an invalid type to use with the key form
of the constructor, if an attempt is used to create
a signatory with the list type, an exception is thrown.