Modifications for TESTNET Compatibility
Modifications to bring the SDK into
compatibility with the testnet updates
of January 23, 2020.  These are breaking
changes, which include:

Changed signature of CallContractAsync
to return a TransactionReceipt. The network
no longer returns the contract ID with the
results of the method call.

Removed ContractID from
CallContractRecord which is returned from
CallContractWithRecordAsync. The network no
longer returns the contract ID with the results
of the method call.

Renamed ContractReceipt to
CreateContractReceipt which is returned from
CreateContractAsync.

Renamed ContractRecord to
CreateContractRecord which is returned from
CreateContractWithRecordAsync.

Renamed AccountReceipt to
CreateAccountReceipt which is returned from
CreateAccountAsync.

Renamed AccountRecord to
CreateAccountRecord which is returned from
CreateAccountWithRecordAsync.

Changed signature of UpdateAccountAsync to
return a TransactionReceipt. The network no
longer returns the Account ID with the results
of the method call.

Changed signature of
UpdateAccountWithRecordAsync to return a
TransactionRecord. The network no longer
returns the Account ID with the results of the
method call.

Temporary relaxing of error checking when
retrieving records, which will be restored in
conjunction with the next testnet release cycle.