Fixes #140 Get Fee Schedule
Added an extension method to retrieve the
Network Fee schedule.  The returned values
contain a dictionary mapping the name of
the functionality to a JSON structure representing
the fee parameters.  It was decided that at this
time it was unecessary to create a large type-safe
fee structure object to represent the data for
concerns of major changes due to the youth of the
network fee structure.