Fixes #196 gRCP Error returned after Accepted Tx
It is possible to receive a gRPC error after the
network has accepted a transaction.  The original
logic would retry sending the transaction after
a grpc error resulting in a duplicate transaction
error.

Logic was changed to query the network after
a grpc error for transactions to ensure that
the transaction had not yet been processed by the
network.  If no receipt is found, the transaction
is re-submitted to the network as before.