Fixes #166 Add Metadata to Mirror Exception
Added custom codes to the MirrorException
produced by the mirror client.  The codes are
mapped from the  underlying gRPC codes
returned from the server.  The gRPC codes are
still avaialble from the inner exception for
those who wish to examine the low-level
details of the error.

Error conditions now include:
* Server is unavalable / not responding
* The topic ID sent in does not exist (new)
* The topic ID sent in is not actually
  a topic (new)
* Any other error error gRPC
  conditions as CommunicationError.