Fixes #171 Add RpcError Code
Added a precheck code error
to designate when the error
originated with the GRPC
connection.  If the remote node
is having connectivity issues,
sometimes "Unknown" is returned
by the SDK.  Considering returning
more descriptive error codes matching
the potential codes from gRPC, but this
is a baby step to start.