Fixes #177 Re-Introduce Contract Endorsement Type
For a while the network had depricated the contract
key type from the system.  It has been recently introduced
back and is used to represent contracts without administrative
keys (perhaps among other uses).  When this type is emitted from
the network, it was causing crashes in the .net sdk as an unknown
key type.  This change resolves the issue.