Fixes #178 Add gRPC Error "Unknown" to Retry
Practice has shown that a node can be temporarily
unavailable and return the "Unknown" status when
a connection attempt is made under load. This is
typically a temporary condition and can be
considered the same as "Unavailable". Add to the
internal Retry loop as a valid reason to retry
before throwing an exception that bubbles up to
the public caller.