Remove docfx.csproj
Since we have not yet automated
the documentation generation we do
not necessarily need the convenience
docfx.csproj in our solution.  It is causing
issues on the build server, so we'll remove
it for now and figure out why when we
get closer to an automation approach
to document generation.