Updates for Gossip Network 0.7.0
Version 0.7.0 of the network no longer allows
creation of accounts with no signing keys, added
a check in the API to throw an argument error if
this is attempted.  If it had passed thru, the error
would have been "Bad Encoding", which is more
difficult for the user to discern.

Gas calculations changed so we upped the gas
price guestimates

other regressions were forund and marked as such
in the integration tests to revisit in case this turns out
to be a bug in the network or intended new behavior.