Fixes #185 SignaturePrefixTrimLimit
Added SignaturePrefixTrimLimit property to the client
configuration context.  This value represents the minimum
allowed signature prefix size for a transaction. The default
is 0, which means that the algorithm will use the smallest
possible value for the PubKeyPrefix value in a signature map,
including removing the property from the transaction when only
one signature exists.