Remove Problematic Test
The test, while useful tests the
network behavior more than the library
itself, and it is difficult to find a
stable initial balance to re-create
running out of funds while sending a
large segmented message.  Therefore
it is removed due to random failures
that are not programing defects other
than computing the proper starting
balance.