Fixes #201 Better Error Message for Rec. Retrieval
Adds a small check for the error code returned
when attempting to get a record.  If it turns
out the cause was the network falsifying
required feels, such information is reported
in the exception instead of the generic
"Can't get Record" type of response.

It was deemed better to do this instead of
retrying to retrieve the record because doing
so required a higher processing fee and
subsequently a signature.  Since a signature
request could invoke a GUI callback or other
such approvals, it was deemed that the user
interface should handle this retry explicitly
instead to be clear to the user what the
problem is.  This is better than the framework
randomly asking for an additional signature
without explanation.