Commissions API Design v 2
With the addition of the Royalty Custom Fee type
to the HAPI, the previous design of the Commissions
requied rework.  Since there are now three distinct
types of fees that can be associated with a token, it
made more sense to consolidate fees into a single
list of ICommission instead of having three separate
lists for each type of commission.

It is not anticipated this will practically add extra levels of
indirection for "setting" types of API calls (in fact things
ought to be simplified), however it may cause more code
to process the results returned from "token info query" calls.

Changes also include various tweaks to integration
tests as the behavior of the network has slightly changed
and returns different error codes for different edge cases.