Add Support for Assess as Surcharge
Added support in the Fractional Commission
for the "Assess as Surcharge" flag.  When set to
true,  the commission fee is applied as an extra
surcharge paid by the  sender of the target token.
If false, the default, the amount of token received
by the reciver is reduced by the fee from the
total amount sent by the sender.