Fixes #190 Accidental Resubmit of Query Payment
Under certain conditions the network will accept a query
with payment and close the connection before returning
results.  The SDK was comingling this edge case with the
edge case where the connection was closed before
submitting.  The re-submission of the query was causing
a duplicate transaction exception.  Added code to recognize
this edge case and report a RPC error instead.

When this happens, unfortunately, the query payment is
lost, the sdk can not re-submit the query message because
the payment has already been processed.  This is not a
defect in the SDK, but of the behavior of the network.