Fixes #204 Make Tests More Robust
Added test to fixtures during setup when
they receive the "change in cost" record
retrival idiosyncry of network behavior to
try to retrieve the record a second time.

This has been added because too many tests
have been failing when the network changes
fee costs between the time of an ask and a
query.  It is not a perfect solution, but the .net
sdk itself should not retry this fetch because doing
so would require an additional signature request
which would be unexpected by calling code.