Updates to Administrative API Calls
Updated the methods involved with
administrative maintnance of the network.
The model by which the network is
administered via upgrade files has changed
necessitating a revision in the API surface.
This is a breaking change, however, since it
is unlikely to affect all but a handfull of
downstream consumers, this will not, at
this time, trigger a major version update.

Also updated some integration tests as a
result of other network behavior changes
and HAPI deprications.