Update Contract Endorsment Implementation
Updated the Endorsement implementation
to represent a "contract key" internally as an
contract's address object.  This is more
consistent with how the key type will be used.
The prior implementation assumed the key was
encoded in ABI form, which experience has shown
adds uneeded complexity.

Also changed token `CreateTokenParams` property
`RoyaltyEndorsement` to `RoyaltiesEndorsement`
for consistency with `CreateAssetParams`.