Address Types Rework
Removed AddressOrAlias due to growing
complexity issues making the API difficult
to understand and explain.  The catalyst for
this was the introduction of the EIP-1014
contract address type.  The discriminated
union approach resulted in an exponential
amount of rework introduced by this
additional dimension.

Removed AddressOrAlias class completely
from the code base.  Revised the Address
account such that it can hold the Alias
or newly introduced Moniker internally
and orcestrates the translation and
switching of types necessary when
converting too and from protobuf messages.

This implementation will be much more
forward compatible with the 8.x implementation
and for now, produces the absolute minimal
API surface to accomplsih the task.  The
Address class introduces two public TryXXX
methods that attempt to extract Alias and
Monikers from the address respectievely,
otherwise the consuming code should rely on
implicit operators to cast a Moniker/Alias
directly to an Address.  If more query methods
are  required to be added to the Address
implementation, they can be added in a
compatible way in the future.