Intermediate work supporting Allowances
Updated implementation supporting recent
changes in protobuf design surrounding
spending allowances.

Necessitated a notable breaking change in the
advanced form of Transfer.  The Transfer Params
now expects a list of CryptoTransfer objects instead
of a dictionary of address->long values.  This is
necessary because "allowance spends" requrie the
delegated/isAllowance flag to be set on the
attempted outflow of the hBar/Crypto value.

Spending Allowances is only supported by the
advanced (TransferParams) transfer methods, all
other convenience methods generate transfers
that assume the owner of the the crypto/token/asset
is explicitly signing the transaciton.