Initial Baby Steps for Mirror Node Library
Introduced a helper library that may or may not
grow into a mirror rest API client.  With this
commit the library is purly experimental and
intened to help identify areas of the greater
SDK that lack will require changes to support
a larger well thought thru SDK architecture.

For example, the SDK does not presently
support JSON conversion from 0.0.123 address
form to a native Address object.  This effort
will ultimately help disover where and how this
support should be implemented.