Custom gRPC Channel Factory
Added the ability to customize how gRPC
Channels are created.  The Gossip & Mirror
node have an additional constructor that
acceps a function to produce a gRPC channel
given the Gateway (for the main client) and
URL (for the mirror client).

This method must be provided when constructing
a root instance of the client.  All cloned clients
will utilize this method when creating new
channels and the channels are shared amongst
cloned clients.