Have to do something about testnet
Added a QueryTip property to the context
so that the user can plus-up a query payment
since the testnet has a tendency of lying to
the user on how much a query will cost.  This
causes integration tests to regularly fail over
the matter and its time to do something
about it.